requirements.txt
yt-dlp>=2024.1.1
anthropic>=0.20.0
google-cloud-storage>=2.14.0
openai>=1.12.0
pyyaml>=6.0.1
requests>=2.31.0
python-dotenv>=1.0.0
